+++ src/felupe/tools/_save.py
@@ -59,7 +59,9 @@ def save(
     >>> region = fem.RegionHexahedron(mesh)
     >>> field = fem.FieldContainer([fem.Field(region, dim=3)])
     >>>
-    >>> boundaries = fem.dof.uniaxial(field, clamped=True, move=0.3)
+    >>> boundaries = fem.dof.uniaxial(
+    ...     field, clamped=True, move=0.3, return_loadcase=False
+    ... )
     >>>
     >>> umat = fem.NeoHooke(mu=1)
     >>> solid = fem.SolidBodyNearlyIncompressible(umat, field, bulk=5000)